Optimizing plugin performance and limiting needed remote calls
Fixes #522

ChangeLog:
  - Updating plugin documentation
  - Updating context Builder
