+++ internal/api/extended.go
@@ -2,32 +2,22 @@ package api
 
 import (
 	"crypto/tls"
-	"log/slog"
 	"net/http"
 
-	"github.com/esnet/gdg/internal/config/domain"
-	"github.com/esnet/gdg/pkg/plugins/secure"
-	"github.com/esnet/gdg/pkg/plugins/secure/contract"
-
 	"github.com/carlmjohnson/requests"
+	"github.com/esnet/gdg/internal/config/domain"
 )
 
 // ExtendedApi provides API request building for Grafana with optional debug mode.
 type ExtendedApi struct {
-	appCfg  *domain.GDGAppConfiguration
-	debug   bool
-	encoder contract.CipherEncoder
+	appCfg *domain.GDGAppConfiguration
+	debug  bool
 }
 
 func NewExtendedApi(cfg *domain.GDGAppConfiguration) *ExtendedApi {
 	o := ExtendedApi{
 		appCfg: cfg,
 		debug:  cfg.IsApiDebug(),
-	}
-	if !cfg.PluginConfig.Disabled && cfg.PluginConfig.CipherPlugin != nil {
-		o.encoder = secure.NewPluginCipherEncoder(cfg.PluginConfig.CipherPlugin, cfg.SecureConfig)
-	} else {
-		o.encoder = secure.NoOpEncoder{}
 	}
 	return &o
 }
@@ -43,21 +33,9 @@ func (extended *ExtendedApi) getRequestBuilder() *requests.Builder {
 	token := extended.appCfg.GetDefaultGrafanaConfig().GetAPIToken()
 
 	if token != "" {
-		decodedValue, err := extended.encoder.DecodeValue(token)
-		if err != nil {
-			slog.Warn("Unable to decode Token using cipher plugin, trying string value")
-		} else {
-			token = decodedValue
-		}
 		req.Header("Authorization", "Bearer "+token)
 	} else {
 		password := extended.appCfg.GetDefaultGrafanaConfig().GetPassword()
-		decodedValue, err := extended.encoder.DecodeValue(password)
-		if err != nil {
-			slog.Warn("Unable to decode Token using cipher plugin, trying string value")
-		} else {
-			password = decodedValue
-		}
 		req.BasicAuth(extended.appCfg.GetDefaultGrafanaConfig().UserName, password)
 	}
