+++ internal/service/login.go
@@ -19,6 +19,10 @@ import (
 // Login sets admin flag and provisions the Extended API for calls unsupported by the OpenAPI spec.
 func (s *DashNGoImpl) Login() {
 	var err error
+	if !s.gdgConfig.PluginConfig.Disabled && s.gdgConfig.PluginConfig.CipherPlugin != nil {
+		s.grafanaConf.UpdateSecureModel(s.encoder.DecodeValue)
+	}
+
 	// Will only succeed for BasicAuth
 	if s.grafanaConf.IsBasicAuth() {
 		var userInfo *models.UserProfileDTO
@@ -97,15 +101,8 @@ func (s *DashNGoImpl) getNewClient(opts ...NewClientOpts) (*client.GrafanaHTTPAP
 // GetClient Returns a new defaultClient given token precedence over Basic Auth
 func (s *DashNGoImpl) GetClient() *client.GrafanaHTTPAPI {
 	if s.grafanaConf.GetAPIToken() != "" {
-		token := s.grafanaConf.GetAPIToken()
-		newToken, err := s.encoder.DecodeValue(token)
-		if err != nil {
-			slog.Warn("unable to decode token", slog.Any("err", err))
-		} else {
-			token = newToken
-		}
 		grafanaClient, _ := s.getNewClient(func(clientCfg *client.TransportConfig) {
-			clientCfg.APIKey = token
+			clientCfg.APIKey = s.grafanaConf.GetAPIToken()
 			clientCfg.Debug = s.GetGlobals().ApiDebug
 		})
 		return grafanaClient
@@ -130,16 +127,8 @@ func (s *DashNGoImpl) GetAdminClient() *client.GrafanaHTTPAPI {
 }
 
 func (s *DashNGoImpl) getDefaultBasicOpts() []NewClientOpts {
-	pass := s.grafanaConf.GetPassword()
-	newPass, err := s.encoder.DecodeValue(pass)
-	if err != nil {
-		slog.Warn("Unable to decode password, falling back on string value", slog.String("password", pass))
-	} else {
-		pass = newPass
-	}
-
 	return []NewClientOpts{func(clientCfg *client.TransportConfig) {
-		clientCfg.BasicAuth = url.UserPassword(s.grafanaConf.UserName, pass)
+		clientCfg.BasicAuth = url.UserPassword(s.grafanaConf.UserName, s.grafanaConf.GetPassword())
 		clientCfg.Debug = s.GetGlobals().ApiDebug
 	}}
 }
