+++ website/content/docs/releases/gdg_0.4.md
@@ -5,7 +5,7 @@ date: 2023-03-31T15:21:01+02:00
 lastmod: 2023-07-13T00:00:00
 draft: false
 images: []
-weight: 4
+weight: 6
 toc: true
 ---
